Fixed Properties and started making query methods.
Originally the Properties fields were type for Dictionaries,
but no objects were being created on init. Now the
Dictionaries are being created, so the objects have a
.Properties.Add() method available.

Started adding the V and has methods. I'm going to try
using Gremlin for queries, and will implement a bit at
a time.